pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dflow"
version = "0.1.0"
description = "Dispersive Flow - A PyTorch library for diffusion models with dispersive loss"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "dFlow Team", email = "contact@dflow.ai"}
]
maintainers = [
    {name = "dFlow Team", email = "contact@dflow.ai"}
]
keywords = [
    "diffusion-models",
    "generative-models", 
    "pytorch",
    "transformers",
    "machine-learning",
    "computer-vision"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "timm>=0.6.0",
    "numpy>=1.21.0",
    "pillow>=8.0.0",
    "tqdm>=4.60.0",
    "wandb>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8", 
    "mypy",
    "pre-commit",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "myst-parser",
    "sphinx-autodoc-typehints",
]
all = [
    "dflow[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/raywang4/DispLoss"
Documentation = "https://dflow.readthedocs.io"
Repository = "https://github.com/raywang4/DispLoss.git"
"Bug Tracker" = "https://github.com/raywang4/DispLoss/issues"

[project.scripts]
dflow-train = "dflow.scripts.train:main"
dflow-sample = "dflow.scripts.sample:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dflow*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.setuptools.package-data]
dflow = ["configs/**/*.py", "configs/**/*.yaml", "configs/**/*.json"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=dflow",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "integration: marks integration tests",
]
